+++ src/app/(pages)/(albaform)/alba/[formId]/edit/page.tsx
@@ -10,9 +10,9 @@ import Button from "@/app/components/button/default/Button";
 import { toast } from "react-hot-toast";
 import { useMutation } from "@tanstack/react-query";
 import { useUpdateProfile } from "@/hooks/queries/user/me/useUpdateProfile";
-import RecruitContentSection from "../../../addform/RecruitContentSection";
-import RecruitConditionSection from "../../../addform/RecruitConditionSection";
-import WorkConditionSection from "../../../addform/WorkConditionSection";
+import RecruitContentSection from "../../../addform/section/RecruitContentSection";
+import RecruitConditionSection from "../../../addform/section/RecruitConditionSection";
+import WorkConditionSection from "../../../addform/section/WorkConditionSection";
 import useUserFormDetail from "@/hooks/queries/form/userFormDetail";
 import { SubmitFormDataType } from "@/types/addform";
 import useEditing from "@/hooks/useEditing";
